Resolve "Add Landau Damping miniapp"

Closes #96

See merge request OPAL/Libraries/ippl!100

+++ src/Communicate/Buffers.cpp
@@ -37,7 +37,7 @@
 
 namespace ippl {
 
-        void Communicate::setDefaultOverallocation(int factor) {
+        void Communicate::setDefaultOverallocation(double factor) {
             defaultOveralloc_m = factor;
         }
